client / server tables like events
