+++ kit/dapp/src/app/[locale]/(private)/assets/(dashboard)/page.tsx
@@ -82,6 +82,21 @@ export default async function AdminDashboard({
           <TransactionsHistory />
         </Suspense>
       </div>
+      <p className="mt-8 mb-4 font-semibold text-2xl">{t("actions-heading")}</p>
+      <div className="grid grid-cols-1 gap-4 divide-x-0 divide-y lg:grid-cols-2 lg:divide-x lg:divide-y-0 2xl:grid-cols-4">
+        <Suspense fallback={<AssetsSupplySkeleton />}>
+          <AssetsSupply />
+        </Suspense>
+        <Suspense fallback={<AssetActivitySkeleton />}>
+          <AssetActivity />
+        </Suspense>
+        <Suspense fallback={<UsersHistorySkeleton />}>
+          <UsersHistory />
+        </Suspense>
+        <Suspense fallback={<TransactionsHistorySkeleton />}>
+          <TransactionsHistory />
+        </Suspense>
+      </div>
       <p className="mt-8 mb-4 font-semibold text-2xl">
         {t("latest-events-heading")}
       </p>
