+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/actions/page.tsx
@@ -1,7 +1,5 @@
 import { ActionsDropdownTable } from "@/components/blocks/actions-table/actions-dropdown-table";
 import { DataTableSkeleton } from "@/components/blocks/data-table/data-table-skeleton";
-import { getUser } from "@/lib/auth/utils";
-import { getActionsList } from "@/lib/queries/actions/actions-list";
 import type { Locale } from "next-intl";
 import { Suspense } from "react";
 import type { Address } from "viem";
@@ -12,17 +10,11 @@ interface PageProps {
 
 export default async function ActionsPage({ params }: PageProps) {
   const { address } = await params;
-  const user = await getUser();
-  const actions = await getActionsList({
-    targetAddress: address,
-    type: "Admin",
-    userAddress: user.wallet,
-  });
-
   return (
     <Suspense fallback={<DataTableSkeleton />}>
       <ActionsDropdownTable
-        actions={actions}
+        type="Admin"
+        targetAddress={address}
         toolbar={{
           enableToolbar: false,
         }}
