+++ kit/dapp/src/components/blocks/actions-table/actions-table.tsx
@@ -1,30 +1,16 @@
-import type { DataTablePaginationOptions } from "@/components/blocks/data-table/data-table-pagination";
-import type { DataTableToolbarOptions } from "@/components/blocks/data-table/data-table-toolbar";
-import {
-  Card,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import { getUser } from "@/lib/auth/utils";
+import { getActionsList } from "@/lib/queries/actions/actions-list";
 import type {
-  Action,
   ActionStatus,
+  ActionType,
 } from "@/lib/queries/actions/actions-schema";
-import { exhaustiveGuard } from "@/lib/utils/exhaustive-guard";
-import type { LucideIcon } from "lucide-react";
-import {
-  AlarmClockCheck,
-  ArrowBigRightDash,
-  CircleDashed,
-  ListCheck,
-} from "lucide-react";
-import { useTranslations } from "next-intl";
-import { DataTable } from "../data-table/data-table";
-import { Columns } from "./actions-columns";
+import type { DataTablePaginationOptions } from "../data-table/data-table-pagination";
+import type { DataTableToolbarOptions } from "../data-table/data-table-toolbar";
+import { ActionsClientTable } from "./actions-client-table";
 
 interface ActionsTableProps {
   status: ActionStatus;
-  actions: Action[];
+  type: ActionType;
   toolbar?: DataTableToolbarOptions;
   pagination?: DataTablePaginationOptions;
 }
@@ -32,93 +18,25 @@ interface ActionsTableProps {
 /**
  * Component that renders actions in a table
  */
-export function ActionsTable({
+export async function ActionsTable({
   status,
-  actions,
+  type,
   toolbar,
   pagination,
 }: ActionsTableProps) {
-  const t = useTranslations("actions");
-
-  let emptyState;
-  switch (status) {
-    case "PENDING": {
-      emptyState = (
-        <EmptyState
-          icon={ListCheck}
-          title={t("tabs.empty-state.title.pending")}
-          description={t("tabs.empty-state.description.pending")}
-        />
-      );
-      break;
-    }
-    case "UPCOMING": {
-      emptyState = (
-        <EmptyState
-          icon={ArrowBigRightDash}
-          title={t("tabs.empty-state.title.upcoming")}
-          description={t("tabs.empty-state.description.upcoming")}
-        />
-      );
-      break;
-    }
-    case "COMPLETED": {
-      emptyState = (
-        <EmptyState
-          icon={CircleDashed}
-          title={t("tabs.empty-state.title.completed")}
-          description={t("tabs.empty-state.description.completed")}
-        />
-      );
-      break;
-    }
-    case "EXPIRED": {
-      emptyState = (
-        <EmptyState
-          icon={AlarmClockCheck}
-          title={t("tabs.empty-state.title.expired")}
-          description={t("tabs.empty-state.description.expired")}
-        />
-      );
-      break;
-    }
-    default:
-      exhaustiveGuard(status);
-  }
-
-  const filteredActions = actions.filter((action) => action.status === status);
-  if (filteredActions.length === 0) {
-    return emptyState;
-  }
+  const user = await getUser();
+  const actionsList = await getActionsList({
+    status,
+    userAddress: user.wallet,
+    type,
+  });
 
   return (
-    <DataTable
-      columns={Columns}
-      columnParams={{ status }}
-      data={filteredActions}
-      name="Actions"
+    <ActionsClientTable
+      status={status}
+      actions={actionsList}
       toolbar={toolbar}
       pagination={pagination}
     />
-  );
-}
-
-interface EmptyStateProps {
-  icon: LucideIcon;
-  title: string;
-  description: string;
-}
-
-function EmptyState({ icon: Icon, title, description }: EmptyStateProps) {
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2 mb-2">
-          <Icon className="size-5" />
-          <div>{title}</div>
-        </CardTitle>
-        <CardDescription>{description}</CardDescription>
-      </CardHeader>
-    </Card>
   );
 }
